requirements.txt
python-chess>=1.9.4
tensorflow>=2.15.0
numpy>=1.26.0
keras>=3.0.0
pyyaml>=6.0.1
matplotlib>=3.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
chess>1.0.0
tqdm>=4.10.0